Otimização de utilização de memória, lógica de cross validation, precisa melhorar a ordenação dos itens avaliados dos folders gerados para otimizar as predições

+++ Netflix/Components/CrossValidation.h
@@ -2,4 +2,4 @@
 #include "Solver.h"
 #include "Util.h"
 
-double crossValidation(int folds, data_input * input, ISolver &solver);
+float crossValidation(int folds, data_input * input, ISolver &solver);
